Refactoring: Now uses audio source as time base. Much precision, such wow! Reduces audio buffer size to 256 samples, about 5ms sample delay. Adds FFT code from bonzomatic, buffer yet too small, needs cycling.

+++ workbench/src/audiostream.hpp
@@ -1,12 +1,13 @@
 #pragma once
 
 #include <vector>
+#include <cstdint>
 
-extern int audio_samplerate;
+extern uint32_t audio_samplerate;
 
-extern int audio_buffersize;
+extern uint32_t audio_buffersize;
 
-extern int audio_channels;
+extern uint32_t audio_channels;
 
 typedef float sample_t;
 
@@ -14,7 +15,7 @@ class AudioStream
 {
 private:
 	std::vector<sample_t> samples;
-	int channels;
+	uint32_t channels;
 public:
 	explicit AudioStream(int channels = 1);
 	AudioStream(AudioStream const &);
@@ -27,24 +28,24 @@ public:
 
 	//! Resizes this stream for another channel count
 	//! @remarks Clears all data!
-	void SetChannels(int chans);
+	void SetChannels(uint32_t chans);
 
 	//! Resizes this stream for another channel count
 	//! @remarks Clears all data!
 	void SetFormatForStream(AudioStream const & other);
 
-	int GetChannels() const { return this->channels; }
+	uint32_t GetChannels() const { return this->channels; }
 
-    int GetLength() const { return int(this->samples.size()) / this->channels; }
+    uint32_t GetLength() const { return uint32_t(this->samples.size()) / this->channels; }
 
 	void CopyTo(AudioStream & target) const;
 
 	sample_t * GetData() { return this->samples.data(); }
 	sample_t const * GetData() const { return this->samples.data(); }
 
-	sample_t & Sample(int idx, int channel) { return this->samples[idx * this->channels + channel]; }
-	sample_t const & Sample(int idx, int channel) const { return this->samples[idx * this->channels + channel]; }
+	sample_t & Sample(size_t idx, uint32_t channel) { return this->samples[idx * this->channels + channel]; }
+	sample_t const & Sample(size_t idx, uint32_t channel) const { return this->samples[idx * this->channels + channel]; }
 
-	sample_t & operator [] (int idx) { return this->samples[idx]; }
-	sample_t const & operator [] (int idx) const { return this->samples[idx]; }
+	sample_t & operator [] (size_t idx) { return this->samples[idx]; }
+	sample_t const & operator [] (size_t idx) const { return this->samples[idx]; }
 };
